Rename edge to ingestion-edge
fixes #40
