+++ ingestion-edge/.dockerignore
