+++ ingestion-edge/pytest.ini
